+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Block PS button on the critical sections
+- Reword import/export to restore/backup (#32)
 
 ## 0.8.0
 - Change default directories to `ux0:/data/savegames`